Made changes to resolve bug.
The logos were not loading in correctly from the companylist componenet to the companycard componenet.
Determined the issue was related to typing the incoming data for the company from the api as logoURL all caps vs logoUrl only U cap.
For some reason this was passing the type check all the way to the CompanyList component and properly pulling out the logoURL info.
But once the prop was passed to the CompanyCard componenet, the prop as spelled could not longer be recognized and it was assigning undefined to the logoURL prop.
The api does return the info as logoUrl, so it was correct to make this change to match the type alias returned by the API.